# lib/api-client-react/src/generated/api.schemas.ts
/**
 * Generated by orval v8.9.1 🍺
 * Do not edit manually.
 * Api
 * AI Assessment Platform API
 * OpenAPI spec version: 0.1.0
 */
export interface HealthStatus {
  status: string;
}

export interface Candidate {
  id: number;
  name: string;
  email: string;
  role: string;
  yearsOfExperience: number;
  skills: string[];
  /** @nullable */
  linkedinUrl?: string | null;
  /** @nullable */
  cvText?: string | null;
  /** @nullable */
  cvFileName?: string | null;
  profileComplete: boolean;
  createdAt: string;
}

export interface CandidateInput {
  /** @minLength 1 */
  name: string;
  email: string;
  /** @minLength 1 */
  role: string;
  /** @minimum 5 */
  yearsOfExperience: number;
  skills: string[];
  linkedinUrl?: string;
}

export interface CandidateUpdate {
  name?: string;
  role?: string;
  /** @minimum 5 */
  yearsOfExperience?: number;
  skills?: string[];
  linkedinUrl?: string;
  profileComplete?: boolean;
}

export interface CVUpload {
  cvText: string;
  cvFileName: string;
}

export interface Role {
  id: number;
  title: string;
  category: string;
  description: string;
}

export type SessionStatus = typeof SessionStatus[keyof typeof SessionStatus];


export const SessionStatus = {
  pending: 'pending',
  in_progress: 'in_progress',
  completed: 'completed',
  evaluated: 'evaluated',
} as const;

export interface Session {
  id: number;
  candidateId: number;
  roleTitle: string;
  /** @nullable */
  jobDescription?: string | null;
  status: SessionStatus;
  questionsGenerated?: boolean;
  /** @nullable */
  startedAt?: string | null;
  /** @nullable */
  completedAt?: string | null;
  createdAt: string;
}

export interface SessionInput {
  candidateId: number;
  roleTitle: string;
  jobDescription?: string;
}

export type SessionUpdateStatus = typeof SessionUpdateStatus[keyof typeof SessionUpdateStatus];


export const SessionUpdateStatus = {
  pending: 'pending',
  in_progress: 'in_progress',
  completed: 'completed',
  evaluated: 'evaluated',
} as const;

export interface SessionUpdate {
  status?: SessionUpdateStatus;
  startedAt?: string;
  completedAt?: string;
}

export type QuestionQuestionType = typeof QuestionQuestionType[keyof typeof QuestionQuestionType];


export const QuestionQuestionType = {
  technical: 'technical',
  system_design: 'system_design',
  behavioral: 'behavioral',
} as const;

export interface Question {
  id: number;
  sessionId: number;
  questionText: string;
  questionType: QuestionQuestionType;
  orderIndex: number;
}

export interface Answer {
  id: number;
  sessionId: number;
  questionId: number;
  transcript: string;
  score: number;
  feedback: string;
  /** @nullable */
  strengths?: string | null;
  /** @nullable */
  weaknesses?: string | null;
  createdAt: string;
}

export interface AnswerInput {
  questionId: number;
  transcript: string;
}

export type EvaluationRating = typeof EvaluationRating[keyof typeof EvaluationRating];


export const EvaluationRating = {
  strong_hire: 'strong_hire',
  hire: 'hire',
  no_hire: 'no_hire',
} as const;

export interface Evaluation {
  id: number;
  sessionId: number;
  overallScore: number;
  roleFitScore: number;
  rating: EvaluationRating;
  technicalScore: number;
  communicationScore: number;
  domainScore: number;
  strengths: string;
  weaknesses: string;
  suggestions: string;
  readyForHiring: boolean;
  /** @nullable */
  summary?: string | null;
  createdAt: string;
}

export interface DashboardStats {
  totalCandidates: number;
  completedSessions: number;
  averageScore: number;
  hireRate: number;
  strongHireCount: number;
  hireCount: number;
  noHireCount: number;
}

export interface SessionSummary {
  id: number;
  candidateName: string;
  roleTitle: string;
  status: string;
  /** @nullable */
  overallScore: number | null;
  /** @nullable */
  rating: string | null;
  createdAt: string;
}

export type ListSessionsParams = {
candidateId?: number;
};