# lib/api-spec/openapi.yaml
openapi: 3.1.0
info:
  title: Api
  version: 0.1.0
  description: AI Assessment Platform API
servers:
  - url: /api
    description: Base API path
...
paths:
...
  /candidates:
    post:
      operationId: createCandidate
      tags: [candidates]
      summary: Create candidate profile
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: "#/components/schemas/CandidateInput"
      responses:
        "201":
          description: Created
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/Candidate"
        "400":
          description: Bad request
  /candidates/{id}:
    get:
      operationId: getCandidate
      tags: [candidates]
      summary: Get candidate profile
      parameters:
        - name: id
          in: path
          required: true
          schema:
            type: integer
      responses:
        "200":
          description: OK
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/Candidate"
        "404":
          description: Not found
    patch:
      operationId: updateCandidate
      tags: [candidates]
      summary: Update candidate profile
      parameters:
        - name: id
          in: path
          required: true
          schema:
            type: integer
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: "#/components/schemas/CandidateUpdate"
      responses:
        "200":
          description: Updated
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/Candidate"
  /candidates/{id}/cv:
    post:
      operationId: uploadCandidateCV
      tags: [candidates]
      summary: Upload CV for candidate
      parameters:
        - name: id
          in: path
          required: true
          schema:
            type: integer
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: "#/components/schemas/CVUpload"
      responses:
        "200":
          description: CV uploaded and parsed
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/Candidate"
...
  /evaluations/{sessionId}:
    get:
      operationId: getEvaluation
      tags: [evaluations]
      summary: Get evaluation result for a session
      parameters:
        - name: sessionId
          in: path
          required: true
          schema:
            type: integer
      responses:
        "200":
          description: OK
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/Evaluation"
        "404":
          description: Not found
...
components:
  schemas:
...
    Candidate:
      type: object
      required: [id, name, email, role, yearsOfExperience, skills, profileComplete, createdAt]
      properties:
        id:
          type: integer
        name:
          type: string
        email:
          type: string
        role:
          type: string
        yearsOfExperience:
          type: integer
        skills:
          type: array
          items:
            type: string
        linkedinUrl:
          type: ["string", "null"]
        cvText:
          type: ["string", "null"]
        cvFileName:
          type: ["string", "null"]
        profileComplete:
          type: boolean
        createdAt:
          type: string
    CandidateInput:
      type: object
      required: [name, email, role, yearsOfExperience, skills]
      properties:
        name:
          type: string
          minLength: 1
        email:
          type: string
        role:
          type: string
          minLength: 1
        yearsOfExperience:
          type: integer
          minimum: 5
        skills:
          type: array
          items:
            type: string
        linkedinUrl:
          type: string
    CandidateUpdate:
      type: object
      properties:
        name:
          type: string
        role:
          type: string
        yearsOfExperience:
          type: integer
          minimum: 5
        skills:
          type: array
          items:
            type: string
        linkedinUrl:
          type: string
        profileComplete:
          type: boolean
    CVUpload:
      type: object
      required: [cvText, cvFileName]
      properties:
        cvText:
          type: string
        cvFileName:
          type: string
...
    Evaluation:
      type: object
      required: [id, sessionId, overallScore, roleFitScore, rating, technicalScore, communicationScore, domainScore, strengths, weaknesses, suggestions, readyForHiring, createdAt]
      properties:
        id:
          type: integer
        sessionId:
          type: integer
        overallScore:
          type: integer
        roleFitScore:
          type: integer
        rating:
          type: string
          enum: [strong_hire, hire, no_hire]
        technicalScore:
          type: integer
        communicationScore:
          type: integer
        domainScore:
          type: integer
        strengths:
          type: string
        weaknesses:
          type: string
        suggestions:
          type: string
        readyForHiring:
          type: boolean
        summary:
          type: ["string", "null"]
        createdAt:
          type: string
...