# lib/api-zod/src/generated/types/candidate.ts
/**
 * Generated by orval v8.9.1 🍺
 * Do not edit manually.
 * Api
 * AI Assessment Platform API
 * OpenAPI spec version: 0.1.0
 */

export interface Candidate {
  id: number;
  name: string;
  email: string;
  role: string;
  yearsOfExperience: number;
  skills: string[];
  /** @nullable */
  linkedinUrl?: string | null;
  /** @nullable */
  cvText?: string | null;
  /** @nullable */
  cvFileName?: string | null;
  profileComplete: boolean;
  createdAt: string;
}