# lib/api-zod/src/generated/types/candidateInput.ts
/**
 * Generated by orval v8.9.1 🍺
 * Do not edit manually.
 * Api
 * AI Assessment Platform API
 * OpenAPI spec version: 0.1.0
 */

export interface CandidateInput {
  /** @minLength 1 */
  name: string;
  email: string;
  /** @minLength 1 */
  role: string;
  /** @minimum 5 */
  yearsOfExperience: number;
  skills: string[];
  linkedinUrl?: string;
}