# lib/api-zod/src/generated/types/candidateUpdate.ts
/**
 * Generated by orval v8.9.1 🍺
 * Do not edit manually.
 * Api
 * AI Assessment Platform API
 * OpenAPI spec version: 0.1.0
 */

export interface CandidateUpdate {
  name?: string;
  role?: string;
  /** @minimum 5 */
  yearsOfExperience?: number;
  skills?: string[];
  linkedinUrl?: string;
  profileComplete?: boolean;
}