# lib/api-zod/src/generated/types/evaluation.ts
/**
 * Generated by orval v8.9.1 🍺
 * Do not edit manually.
 * Api
 * AI Assessment Platform API
 * OpenAPI spec version: 0.1.0
 */
import type { EvaluationRating } from './evaluationRating';

export interface Evaluation {
  id: number;
  sessionId: number;
  overallScore: number;
  roleFitScore: number;
  rating: EvaluationRating;
  technicalScore: number;
  communicationScore: number;
  domainScore: number;
  strengths: string;
  weaknesses: string;
  suggestions: string;
  readyForHiring: boolean;
  /** @nullable */
  summary?: string | null;
  createdAt: string;
}